+++ src/components/Block.tsx
@@ -3,7 +3,6 @@ import classnames from 'classnames'
 import FieldDocLabel from './FieldDocLabel'
 import Doc from './Doc'
 
-
 type BlockProps = PropsWithChildren & {
   "data-wd-key"?: string
   label?: string