+++ src/components/_SpecProperty.tsx
@@ -1,20 +1,21 @@
 import React from 'react'
 
-import SpecField, {SpecFieldProps} from './SpecField'
+import FieldSpec, {FieldSpecProps} from './FieldSpec'
 import FunctionButtons from './_FunctionButtons'
 
 import labelFromFieldName from '../libs/label-from-field-name'
 
 
-type SpecPropertyProps = SpecFieldProps & {
-  onZoomClick(...args: unknown[]): unknown
-  onDataClick(...args: unknown[]): unknown
+type SpecPropertyProps = FieldSpecProps & {
   fieldName?: string
   fieldType?: string
   fieldSpec?: any
   value?: any
   errors?: {[key: string]: {message: string}}
-  onExpressionClick?(...args: unknown[]): unknown
+  onZoomClick(): void
+  onDataClick(): void
+  onExpressionClick?(): void
+  onElevationClick?(): void
 };
 
 
@@ -31,11 +32,12 @@ export default class SpecProperty extends React.Component<SpecPropertyProps> {
       onZoomClick={this.props.onZoomClick}
       onDataClick={this.props.onDataClick}
       onExpressionClick={this.props.onExpressionClick}
+      onElevationClick={this.props.onElevationClick}
     />
 
     const error = errors![fieldType+"."+fieldName as any] as any;
 
-    return <SpecField
+    return <FieldSpec
       {...this.props}
       error={error}
       fieldSpec={this.props.fieldSpec}