+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 ## main
 
 ### ✨ Features and improvements
+- Added translation to "Links" in debug modal
+- Add support for hillshade's color arrays and relief-color elevation expression
 - _...Add new stuff here..._
 
 ### 🐞 Bug fixes
 
+- Fixed an issue when clicking on a popup and then clicking on the map again
+- Fix modal close button possition
+- Fixed an issue with the generation of tranlations
+- Fix missing spec info when clicking next to a property
 - _...Add new stuff here..._
 
 ## 3.0.0